--- a/.env.example
+++ b/.env.example
@@ -2,3 +2,4 @@ HOST=0.0.0.0
 PORT=3000
 DATABASE_URL="sqlite://dev.db"
 # DATABASE_URL="postgresql://postgres:postgres@localhost/clean-axum"
+PREFORK=1