--- a/Cargo.toml
+++ b/Cargo.toml
@@ -37,11 +37,12 @@ tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 # runtime
 axum = { workspace = true, features = ["tokio", "http1", "http2"] }
 tokio = { version = "1.38.0", features = ["full"] }
+prefork = { version = "0.5.1", default-features = false, optional = true }
 
 # shuttle runtime
-shuttle-axum = { version = "0.46.0", optional = true }
-shuttle-runtime = { version = "0.46.0", optional = true }
-shuttle-shared-db = { version = "0.46.0", features = [
+shuttle-axum = { version = "0.47.0", optional = true }
+shuttle-runtime = { version = "0.47.0", optional = true }
+shuttle-shared-db = { version = "0.47.0", features = [
     "postgres",
 ], optional = true }
 
@@ -53,5 +54,6 @@ http-body-util = "0.1.2"
 serde_json = { workspace = true }
 
 [features]
-default = []
+default = ["prefork"]
+prefork = ["prefork/tokio"]
 shuttle = ["shuttle-axum", "shuttle-runtime", "shuttle-shared-db"]