--- a/api/src/init.rs
+++ b/api/src/init.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use axum::Router;
 use sea_orm::{ConnectOptions, Database, DatabaseConnection};
 
@@ -16,8 +18,14 @@ pub fn setup_config() -> Config {
     Config::from_env()
 }
 
-pub async fn setup_db(db_url: &str) -> DatabaseConnection {
-    let opt = ConnectOptions::new(db_url);
+pub async fn setup_db(db_url: &str, prefork: bool) -> DatabaseConnection {
+    let mut opt = ConnectOptions::new(db_url);
+    opt.max_lifetime(Duration::from_secs(60));
+
+    if !prefork {
+        opt.min_connections(10).max_connections(100);
+    }
+
     Database::connect(opt)
         .await
         .expect("Database connection failed")