--- a/src/main.rs
+++ b/src/main.rs
@@ -1,12 +1,9 @@
-use api::{setup_db, setup_router};
-use doc::ApiDoc;
-use utils::migrate;
+#[cfg(not(feature = "shuttle"))]
+mod tokio;
 
 #[cfg(not(feature = "shuttle"))]
-#[tokio::main]
-async fn main() {
+fn main() {
     use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
-    use utils::create_dev_db;
 
     tracing_subscriber::registry()
         .with(
@@ -17,30 +14,14 @@ async fn main() {
         .init();
 
     tracing::info!("Starting with tokio");
-
-    let config = api::setup_config();
-    create_dev_db(&config.db_url);
-
-    let conn = setup_db(&config.db_url).await;
-    migrate(&conn).await.expect("Migration failed!");
-
-    let router = setup_router(conn).attach_doc();
-    let listener = tokio::net::TcpListener::bind(&config.get_server_url())
-        .await
-        .unwrap();
-
-    tracing::debug!("listening on http://{}", listener.local_addr().unwrap());
-    axum::serve(listener, router).await.unwrap();
+    tokio::run();
 }
+
+#[cfg(feature = "shuttle")]
+mod shuttle;
 
 #[cfg(feature = "shuttle")]
 #[shuttle_runtime::main]
 async fn main(#[shuttle_shared_db::Postgres] db_url: String) -> shuttle_axum::ShuttleAxum {
-    tracing::info!("Starting with shuttle");
-
-    let conn = setup_db(&db_url).await;
-    migrate(&conn).await.expect("Migration failed!");
-
-    let router = setup_router(conn).attach_doc();
-    Ok(router.into())
+    shuttle::run(&db_url).await
 }