On Intel i9-10900K

## Before

```bash
$ wrk --latency -t20 -c50 -d10s http://127.0.0.1:3000/users\?username\=
Running 10s test @ http://127.0.0.1:3000/users?username=
  20 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   848.39us    1.97ms  69.61ms   98.86%
    Req/Sec     2.75k   270.89     4.04k    79.21%
  Latency Distribution
     50%  609.00us
     75%    0.92ms
     90%    1.39ms
     99%    2.94ms
  550869 requests in 10.10s, 135.01MB read
Requests/sec:  54542.49
Transfer/sec:     13.37MB
```

## After

```bash
$ wrk --latency -t20 -c50 -d10s http://127.0.0.1:3000/users\?username\=
Running 10s test @ http://127.0.0.1:3000/users?username=
  20 threads and 50 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   356.84us  326.68us  20.21ms   93.21%
    Req/Sec     5.88k     2.37k   12.17k    61.26%
  Latency Distribution
     50%  282.00us
     75%  488.00us
     90%  647.00us
     99%    1.15ms
  1179612 requests in 10.10s, 289.12MB read
Requests/sec: 116806.23
Transfer/sec:     28.63MB
```